# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Test Documentation",
        "description": "Docmentation for CRM Test",
        "contact": {
            "email": "gilaarief@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "CRM Test Server"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user details",
                "operationId": "357c4e3728598d3c5c9887e1ad84e5d5",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile-update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile",
                "operationId": "b7daf2e380b451ff42f3cbe86a624d8f",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Profile's name",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Profile's address",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Profile's phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    },
...
}